// src/app/Models/navegate-by-path.ts
// Generated by https://quicktype.io

export interface INavigateByPathResponse {
  ids:                  number[];
  entries:              Entry[];
  driveNavigationToken: string;
}

export interface Entry {
  mftId:            number;
  length:           number;
  name:             string;
  createdDate:      string;
  modifiedDate:     string;
  lastAccessedDate: string;
  isDirectory:      boolean;
  isHidden:         boolean;
  isSystem:         boolean;
}